Separate file reading from file creation
