+++ lib/pansophy/local.rb
@@ -1,2 +1,3 @@
 require 'pansophy/local/directory'
+require 'pansophy/local/create_file'
 require 'pansophy/local/file'
