+++ lib/pansophy/local/create_file.rb
@@ -1,0 +1,28 @@
+module Pansophy
+  module Local
+    class CreateFile
+      include Adamantium::Flat
+
+      def initialize(path, body)
+        @pathname = Pathname.new(path)
+        @body     = body
+      end
+
+      def call(options = {})
+        prevent_overwrite! unless options[:overwrite]
+        @pathname.dirname.mkpath
+        ::File.open(@pathname, 'w') do |f|
+          f.write @body
+        end
+      end
+
+      private
+
+      def prevent_overwrite!
+        return unless @pathname.exist?
+        fail ArgumentError,
+             "#{@pathname} already exists, pass ':overwrite => true' to overwrite"
+      end
+    end
+  end
+end
