+++ lib/pansophy/local/directory.rb
@@ -19,7 +19,7 @@ module Pansophy
       end
 
       def create_file(path, body, options = {})
-        Local::File.new(pathname.join(path), body).create(options)
+        CreateFile.new(pathname.join(path), body).call(options)
       end
 
       private
